Musicians and Bands Controller updated. Names changed. New migration.

+++ MusicManager.API/Features/Bands/BandsController.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Mvc;
 using MusicManager.API.Common.CustomActionFilters;
 using MusicManager.API.Data.Models;
 using MusicManager.API.Features.Bands.Models;
+using MusicManager.API.Features.Users;
 
 namespace MusicManager.API.Features.Bands
 {
@@ -11,14 +12,15 @@ namespace MusicManager.API.Features.Bands
     [ApiController]
     public class BandsController : ControllerBase
     {
-        private readonly BandRepository bandRepository;
+        private readonly BandService bandRepository;
         private readonly IMapper mapper;
+        private readonly IUserService userService;
 
-
-        public BandsController(BandRepository bandRepository, IMapper mapper)
+        public BandsController(BandService bandRepository, IMapper mapper, IUserService userService)
         {
             this.bandRepository = bandRepository;
             this.mapper = mapper;
+            this.userService = userService;
         }
 
         [HttpGet]
@@ -26,7 +28,7 @@ namespace MusicManager.API.Features.Bands
         {
             var bands = await bandRepository.AllAsync();
 
-            var bandDtos = mapper.Map<List<BandSingleDto>>(bands);
+            var bandDtos = mapper.Map<List<BandDetails>>(bands);
 
             return Ok(bandDtos);
         }
@@ -42,48 +44,61 @@ namespace MusicManager.API.Features.Bands
                 return NotFound();
             }
 
-            var bandDto = mapper.Map<BandSingleDto>(band);
+            var bandDto = mapper.Map<BandDetails>(band);
 
             return Ok(bandDto);
         }
 
         [HttpPost]
-        [ValidateModel]
         [Authorize]
-        public async Task<IActionResult> CreateAsync([FromBody] CreateBandRequestDto createBandRequestDto)
+        public async Task<IActionResult> CreateAsync([FromBody] CreateBandRequestModel createBandRequestDto)
         {
             var band = mapper.Map<Band>(createBandRequestDto);
+            var user = this.userService.GetCurrentUserDetails().Result;
+            if (user == null)
+            {
+                return BadRequest("Not authorized!");
+            }
 
+            band.CreatedBy = user.Email;
+            band.UserId = user.Id;
             band = await bandRepository.CreateAsync(band);
 
-            var bandDto = mapper.Map<BandDto>(band);
+            var bandDto = mapper.Map<BandModel>(band);
 
-            return CreatedAtAction(nameof(GetById), new { id = bandDto.Id }, bandDto);
+            return Ok(bandDto);
         }
 
         [HttpPut]
         [ValidateModel]
         [Route("{id:int}")]
-        [Authorize(Roles = "Admin,Creator")]
-        public async Task<IActionResult> UpdateAsync([FromRoute] int id, [FromBody] UpdateBandRequestDto updateBandRequestDto)
+        [Authorize]
+        public async Task<IActionResult> UpdateAsync([FromRoute] int id, [FromBody] UpdateBandRequestModel updateBandRequestDto)
         {
             var band = await bandRepository.ByIdAsync(id);
-
             if (band == null)
             {
                 return NotFound();
             }
 
+            var user = this.userService.GetCurrentUserDetails().Result;
+            if (user.Email != band.CreatedBy)
+            {
+                return BadRequest("Not authorized!");
+            }
+
+            band.ModifiedBy = user.Email;
             band.Name = updateBandRequestDto.Name;
             band.Style = updateBandRequestDto.Style;
 
             bandRepository.Update(band);
 
-            return Ok(mapper.Map<BandDto>(band));
+            return Ok(mapper.Map<BandModel>(band));
         }
 
         [HttpDelete]
         [Route("{id:int}")]
+        [Authorize]
         public async Task<IActionResult> DeleteAsync([FromRoute] int id)
         {
             var band = await bandRepository.ByIdAsync(id);
@@ -93,9 +108,17 @@ namespace MusicManager.API.Features.Bands
                 return NotFound();
             }
 
+            var user = this.userService.GetCurrentUserDetails().Result;
+            if (user.Email != band.CreatedBy)
+            {
+                return BadRequest("Not authorized!");
+            }
+
+            band.DeletedBy = user.Email;
+
             bandRepository.Delete(band);
 
-            return Ok(mapper.Map<BandDto>(band));
+            return Ok(mapper.Map<BandModel>(band));
         }
     }
 }
