+++ MusicManager.API/Features/Musicians/MusiciansController.cs
@@ -1,8 +1,12 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using MusicManager.API.Common.CustomActionFilters;
+using MusicManager.API.Common.Models;
 using MusicManager.API.Data.Models;
+using MusicManager.API.Features.Bands;
 using MusicManager.API.Features.Musicians.Models;
+using MusicManager.API.Features.Users;
 
 namespace MusicManager.API.Features.Musicians
 {
@@ -10,13 +14,17 @@ namespace MusicManager.API.Features.Musicians
     [ApiController]
     public class MusiciansController : ControllerBase
     {
-        private readonly MusicianRepository musicianRepository;
+        private readonly MusicianService musicianRepository;
         private readonly IMapper mapper;
+        private readonly IUserService userService;
+        private readonly BandService bandService;
 
-        public MusiciansController(MusicianRepository musicianRepository, IMapper mapper)
+        public MusiciansController(MusicianService musicianRepository, IMapper mapper, IUserService userService, BandService bandService)
         {
             this.musicianRepository = musicianRepository;
             this.mapper = mapper;
+            this.userService = userService;
+            this.bandService = bandService;
         }
 
         [HttpGet]
@@ -24,7 +32,7 @@ namespace MusicManager.API.Features.Musicians
         {
             var musicians = await musicianRepository.AllAsync();
 
-            var musicianDtos = mapper.Map<List<MusicianDto>>(musicians);
+            var musicianDtos = mapper.Map<List<MusicianModel>>(musicians);
 
             return Ok(musicianDtos);
         }
@@ -40,33 +48,49 @@ namespace MusicManager.API.Features.Musicians
                 return NotFound();
             }
 
-            var musicianDto = mapper.Map<MusicianDto>(musician);
+            var musicianDto = mapper.Map<MusicianModel>(musician);
 
             return Ok(musicianDto);
         }
 
         [HttpPost]
         [ValidateModel]
-        public async Task<IActionResult> CreateAsync([FromBody] CreateMusicianRequestDto createBandRequestDto)
+        [Authorize]
+        public async Task<IActionResult> CreateAsync([FromBody] CreateMusicianRequestModel createBandRequestDto)
         {
-            var musician = mapper.Map<Musician>(createBandRequestDto);
-
-            await musicianRepository.CreateAsync(musician);
-
-            if (musician.Band == null)
+            var user = this.userService.GetCurrentUserDetails().Result;
+            if (user == null)
             {
-                return NotFound();
+                return BadRequest("Not authorized!");
             }
 
-            var musicianDto = mapper.Map<MusicianDto>(musician);
+            var band = bandService.ByIdAsync(createBandRequestDto.BandId).Result;
+            if (band == null)
+            {
+                return BadRequest($"There is not band with id: {createBandRequestDto.BandId}");
+            }
+
+            if (band.CreatedBy != user.Email)
+            {
+                return BadRequest("Not authorized!");
+            }
+
+            var musician = mapper.Map<Musician>(createBandRequestDto);
+            musician.CreatedBy = user.Email;
+            musician.BandId = band.Id;
+            band.Musicians.Add(musician);
+            await musicianRepository.CreateAsync(musician);
+
+            var musicianDto = mapper.Map<MusicianModel>(musician);
 
             return Ok(musicianDto);
         }
 
         [HttpPut]
         [ValidateModel]
+        [Authorize]
         [Route("{id:Guid}")]
-        public async Task<IActionResult> UpdateAsync([FromRoute] Guid id, [FromBody] UpdateMusicianRequestDto updateBandRequestDto)
+        public async Task<IActionResult> UpdateAsync([FromRoute] Guid id, [FromBody] UpdateMusicianRequestModel updateBandRequestDto)
         {
             var musician = await musicianRepository.ByIdAsync(id);
 
@@ -75,15 +99,24 @@ namespace MusicManager.API.Features.Musicians
                 return NotFound();
             }
 
+            var user = this.userService.GetCurrentUserDetails().Result;
+            if (user == null)
+            {
+                return BadRequest("Not authorized!");
+            }
+
+            musician.CreatedBy = user.Email;
+
             musician.Name = updateBandRequestDto.Name;
             musician.Clothing = updateBandRequestDto.Clothing;
 
             musicianRepository.Update(musician);
 
-            return Ok(mapper.Map<MusicianDto>(musician));
+            return Ok(mapper.Map<MusicianModel>(musician));
         }
 
         [HttpDelete]
+        [Authorize]
         [Route("{id:Guid}")]
         public async Task<IActionResult> DeleteAsync([FromRoute] Guid id)
         {
@@ -94,9 +127,17 @@ namespace MusicManager.API.Features.Musicians
                 return NotFound();
             }
 
+            var user = this.userService.GetCurrentUserDetails().Result;
+            if (user == null)
+            {
+                return BadRequest("Not authorized!");
+            }
+
+            musician.CreatedBy = user.Email;
+
             musicianRepository.Delete(musician);
 
-            return Ok(mapper.Map<MusicianDto>(musician));
+            return Ok(mapper.Map<MusicianModel>(musician));
         }
     }
 }
